# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservice para gestão de cobranças (inicialmente integração Asaas; no futuro bancos diversos).",
        "title": "Charges Service",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8083",
    "basePath": "/",
    "paths": {
...
        "/v1/asaas/charges/{id}": {
            "delete": {
                "description": "Exclui uma cobrança (payment) do Asaas e remove do banco de dados iam.charges",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asaas Charges"
                ],
                "summary": "Excluir cobrança do Asaas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da cobrança no Asaas (pay_xxx)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do escritório contábil",
                        "name": "accounting_office_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cobrança excluída com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Cobrança não encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
}